requirements.txt
pyyaml>=6.0
requests>=2.31.0
certifi>=2024.2.2
